rust: merge parser.rs into applayer.rs
Both were serving the same purpose.
