+++ rust/src/lib.rs
@@ -48,8 +48,6 @@ pub mod json;
 pub mod applayer;
 pub mod filecontainer;
 pub mod filetracker;
-#[macro_use]
-pub mod parser;
 pub mod kerberos;
 
 #[cfg(feature = "lua")]
